ARC: int vs long
- arch specific bitsperlong.h
- general LP64 related changes
- don't force cast argument to __builtin_arc_sr()
- irqflags: STAT32 is 32-bits but generic raw_local_irq_* expect flags to be long

+++ arch/arc/include/asm/processor.h
@@ -67,7 +67,7 @@ struct task_struct;
 extern void start_thread(struct pt_regs * regs, unsigned long pc,
 			 unsigned long usp);
 
-extern unsigned int __get_wchan(struct task_struct *p);
+extern unsigned long __get_wchan(struct task_struct *p);
 
 #endif /* !__ASSEMBLY__ */
